# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API Handle Dentists, Patients and Appointments",
        "title": "Proyecto Final Go",
        "contact": {
            "name": "Melania Simes and Laura Urrego",
            "url": "https://github.com/laura-urrego-valois/desafio-final-go.git"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/dentists": {
            "get": {
                "description": "This endpoint allows you to retrieve all patients.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get all patients",
                "responses": {
                    "200": {
                        "description": "Patients",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Patient"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve patients"
                    }
                }
            },
            "put": {
                "description": "This endpoint allows you to update a dentist with the provided data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentists"
                ],
                "summary": "Update a dentist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Updated dentist information",
                        "name": "dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dentist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated dentist",
                        "schema": {
                            "$ref": "#/definitions/domain.Dentist"
                        }
                    },
                    "400": {
                        "description": "Invalid dentist data or missing required fields"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "500": {
                        "description": "Failed to update dentist"
                    }
                }
            },
            "post": {
                "description": "This endpoint allows you to create a new dentist with the provided data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentists"
                ],
                "summary": "Create a new dentist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dentist",
                        "name": "dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Dentist"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Dentist created successfully"
                    },
                    "400": {
                        "description": "Invalid dentist data or missing required fields"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "500": {
                        "description": "Failed to create dentist"
                    }
                }
            }
        },
...
        "/patients": {
            "put": {
                "description": "This endpoint allows you to update a patient with the provided data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Updated patient information",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated patient",
                        "schema": {
                            "$ref": "#/definitions/domain.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid patient data or missing required fields"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "500": {
                        "description": "Failed to update patient"
                    }
                }
            },
            "post": {
                "description": "This endpoint allows you to create a new patient with the provided data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Patient",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Patient"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Patient created successfully"
                    },
                    "400": {
                        "description": "Invalid patient data or missing required fields"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "500": {
                        "description": "Failed to create patient"
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "description": "This endpoint allows you to retrieve a patient by their ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get a patient by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient",
                        "schema": {
                            "$ref": "#/definitions/domain.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            },
            "delete": {
                "description": "This endpoint allows you to delete a patient by their ID.",
                "tags": [
                    "Patients"
                ],
                "summary": "Delete a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient deleted successfully"
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "500": {
                        "description": "Failed to delete patient"
                    }
                }
            },
            "patch": {
                "description": "This endpoint allows you to update a patient's address with the provided data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient's address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated address information",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address updated successfully"
                    },
                    "400": {
                        "description": "Invalid request or missing required fields"
                    },
                    "401": {
                        "description": "Unauthorized access due to missing or invalid token"
                    },
                    "404": {
                        "description": "Patient not found"
                    },
                    "500": {
                        "description": "Failed to update Address"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Dentist": {
            "type": "object",
            "required": [
                "FirstName",
                "LastName",
                "License"
            ],
            "properties": {
                "FirstName": {
                    "description": "@Description The first name of the dentist\n@Example \"Daniel\"",
                    "type": "string"
                },
                "Id": {
                    "description": "@Description The unique identifier of the dentist\n@Example 1",
                    "type": "integer"
                },
                "LastName": {
                    "description": "@Description The last name of the dentist\n@Example \"Rodríguez\"",
                    "type": "string"
                },
                "License": {
                    "description": "@Description The license number of the dentist\n@Example \"AXMER\"",
                    "type": "string"
                }
            }
        },
        "domain.Patient": {
            "type": "object",
            "required": [
                "Address",
                "DNI",
                "FirstName",
                "LastName",
                "ReleaseDate"
            ],
            "properties": {
                "Address": {
                    "description": "@Description The address of the patient\n@Example \"Av. 22 # 40\"",
                    "type": "string"
                },
                "DNI": {
                    "description": "@Description The DNI of the patient\n@Example \"538434\"",
                    "type": "string"
                },
                "FirstName": {
                    "description": "@Description The first name of the patient\n@Example \"Daniel\"",
                    "type": "string"
                },
                "Id": {
                    "description": "@Description The unique identifier of the patient\n@Example 1",
                    "type": "integer"
                },
                "LastName": {
                    "description": "@Description The last name of the patient\n@Example \"Rodríguez\"",
                    "type": "string"
                },
                "ReleaseDate": {
                    "description": "@Description The release date of the patient (dd/MM/YYYY)\n@Example \"30/03/2024\"",
                    "type": "string"
                }
            }
        }
    }
}